+++ crates/node/src/sync/blobs.rs
@@ -5,7 +5,6 @@ use calimero_primitives::identity::PublicKey;
 use eyre::bail;
 use futures_util::stream::poll_fn;
 use futures_util::TryStreamExt;
-use libp2p::PeerId;
 use rand::seq::IteratorRandom;
 use rand::thread_rng;
 use tokio::sync::mpsc;
@@ -16,23 +15,16 @@ use crate::types::{InitPayload, MessagePayload, StreamMessage};
 use crate::Node;
 
 impl Node {
-    pub async fn initiate_blob_share_process(
+    pub(super) async fn initiate_blob_share_process(
         &self,
         context: &Context,
+        our_identity: PublicKey,
         blob_id: BlobId,
         size: u64,
-        chosen_peer: PeerId,
+        stream: &mut Stream,
     ) -> eyre::Result<()> {
-        let identities = self.ctx_manager.get_context_owned_identities(context.id)?;
-
-        let Some(our_identity) = identities.into_iter().choose(&mut thread_rng()) else {
-            bail!("no identities found for context: {}", context.id);
-        };
-
-        let mut stream = self.network_client.open_stream(chosen_peer).await?;
-
         send(
-            &mut stream,
+            stream,
             &StreamMessage::Init {
                 context_id: context.id,
                 party_id: our_identity,
@@ -41,7 +33,7 @@ impl Node {
         )
         .await?;
 
-        let Some(ack) = recv(&mut stream, self.sync_config.timeout).await? else {
+        let Some(ack) = recv(stream, self.sync_config.timeout).await? else {
             bail!("no response to blob share request");
         };
 
@@ -82,7 +74,7 @@ impl Node {
         let read_task = async {
             let mut sequencer = Sequencer::default();
 
-            while let Some(msg) = recv(&mut stream, self.sync_config.timeout).await? {
+            while let Some(msg) = recv(stream, self.sync_config.timeout).await? {
                 let (sequence_id, chunk) = match msg {
                     StreamMessage::OpaqueError => bail!("other peer ran into an error"),
                     StreamMessage::Message {
@@ -117,7 +109,7 @@ impl Node {
         Ok(())
     }
 
-    pub async fn handle_blob_share_request(
+    pub(super) async fn handle_blob_share_request(
         &self,
         context: Context,
         their_identity: PublicKey,
