+++ crates/node/src/sync/state.rs
@@ -6,7 +6,6 @@ use calimero_primitives::context::Context;
 use calimero_primitives::hash::Hash;
 use calimero_primitives::identity::PublicKey;
 use eyre::{bail, OptionExt};
-use libp2p::PeerId;
 use rand::seq::IteratorRandom;
 use rand::thread_rng;
 use tracing::debug;
@@ -16,21 +15,14 @@ use crate::types::{InitPayload, MessagePayload, StreamMessage};
 use crate::Node;
 
 impl Node {
-    pub async fn initiate_state_sync_process(
+    pub(super) async fn initiate_state_sync_process(
         &self,
         context: &mut Context,
-        chosen_peer: PeerId,
+        our_identity: PublicKey,
+        stream: &mut Stream,
     ) -> eyre::Result<()> {
-        let identities = self.ctx_manager.get_context_owned_identities(context.id)?;
-
-        let Some(our_identity) = identities.into_iter().choose(&mut thread_rng()) else {
-            bail!("no identities found for context: {}", context.id);
-        };
-
-        let mut stream = self.network_client.open_stream(chosen_peer).await?;
-
         send(
-            &mut stream,
+            stream,
             &StreamMessage::Init {
                 context_id: context.id,
                 party_id: our_identity,
@@ -42,7 +34,7 @@ impl Node {
         )
         .await?;
 
-        let Some(ack) = recv(&mut stream, self.sync_config.timeout).await? else {
+        let Some(ack) = recv(stream, self.sync_config.timeout).await? else {
             bail!("no response to state sync request");
         };
 
@@ -80,7 +72,7 @@ impl Node {
         let mut sqx_out = Sequencer::default();
 
         send(
-            &mut stream,
+            stream,
             &StreamMessage::Message {
                 sequence_id: sqx_out.next(),
                 payload: MessagePayload::StateSync {
@@ -90,19 +82,13 @@ impl Node {
         )
         .await?;
 
-        self.bidirectional_sync(
-            context,
-            our_identity,
-            their_identity,
-            &mut sqx_out,
-            &mut stream,
-        )
-        .await?;
+        self.bidirectional_sync(context, our_identity, their_identity, &mut sqx_out, stream)
+            .await?;
 
         Ok(())
     }
 
-    pub async fn handle_state_sync_request(
+    pub(super) async fn handle_state_sync_request(
         &self,
         context: Context,
         their_identity: PublicKey,
